fix: move data fetching into card line chart to prevent rendering loop
